feat: Add react-hook-form dependency and update package.json
This commit adds the `react-hook-form` dependency to the project by updating the `package.json` file. The new version `7.51.5` is included in the dependencies section. This update is necessary to implement form functionality in the code changes.

The purpose of this commit is to add the `react-hook-form` library to the project and update the package.json file to reflect this change.